requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
attrs==23.2.0
black==24.4.2
click==8.1.7
discord.py==2.3.2
dnspython==2.7.0
frozenlist==1.4.1
idna==3.7
motor==3.6.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
pytz
PyYAML==6.0.2
six==1.16.0
yarl==1.9.4
